+++ src/Type/Definition/InputObjectField.php
@@ -10,7 +10,6 @@ use GraphQL\Type\Schema;
 use GraphQL\Utils\Utils;
 
 use function array_key_exists;
-use function sprintf;
 
 class InputObjectField
 {
@@ -82,27 +81,16 @@ class InputObjectField
             throw new InvariantViolation("{$parentType->name}.{$this->name}: {$error->getMessage()}");
         }
 
-        $type = $this->getType();
-        if ($type instanceof WrappingType) {
-            $type = $type->getWrappedType(true);
+        $type = Type::getNamedType($this->getType());
+
+        if (! $type instanceof InputType) {
+            $notInputType = Utils::printSafe($this->type);
+
+            throw new InvariantViolation("{$parentType->name}.{$this->name} field type must be Input Type but got: {$notInputType}");
         }
 
-        Utils::invariant(
-            $type instanceof InputType,
-            sprintf(
-                '%s.%s field type must be Input Type but got: %s',
-                $parentType->name,
-                $this->name,
-                Utils::printSafe($this->type)
-            )
-        );
-        Utils::invariant(
-            ! array_key_exists('resolve', $this->config),
-            sprintf(
-                '%s.%s field has a resolve property, but Input Types cannot define resolvers.',
-                $parentType->name,
-                $this->name
-            )
-        );
+        if (array_key_exists('resolve', $this->config)) {
+            throw new InvariantViolation("{$parentType->name}.{$this->name} field has a resolve property, but Input Types cannot define resolvers.");
+        }
     }
 }